todo_organizer/users/serializers.py
from rest_framework.serializers import HyperlinkedModelSerializer

from users.models import CustomUser


class UserModelSerializer(HyperlinkedModelSerializer):
    class Meta:
        model = CustomUser
        fields = ['id', 'username', 'first_name', 'last_name', 'email', 'url']
